refactor(redteam): isolate provider context extraction
