+++ src/redteam/remoteGenerationContext.ts
@@ -1,12 +1,8 @@
-import { getProviderIds } from '../providers/index';
 import { getCloudDatabaseId, isCloudProvider } from '../util/cloud';
 
-import type { UnifiedConfig } from '../types/index';
 import type { RedteamGenerationContext } from './types';
 
 export type { RedteamGenerationContext } from './types';
-
-type ProviderConfigLike = Partial<UnifiedConfig>['providers'];
 
 function isRecord(value: unknown): value is Record<string, unknown> {
   return value !== null && typeof value === 'object' && !Array.isArray(value);
@@ -16,16 +12,6 @@ function getCloudTargetIdFromProviderId(providerId: unknown): string | undefined
   return typeof providerId === 'string' && isCloudProvider(providerId)
     ? getCloudDatabaseId(providerId)
     : undefined;
-}
-
-function getLinkedCloudTargetId(value: unknown): string | undefined {
-  if (!isRecord(value)) {
-    return undefined;
-  }
-
-  const config = isRecord(value.config) ? value.config : value;
-  const linkedTargetId = config.linkedTargetId;
-  return getCloudTargetIdFromProviderId(linkedTargetId);
 }
 
 export function remoteGenerationContextPayload(contextOrCloudTargetId?: unknown): {
@@ -49,83 +35,6 @@ export function getCloudTargetIdFromTargetIds(targetIds: string[]): string | und
     }
   }
   return undefined;
-}
-
-export function getProviderTargetIds(providers: ProviderConfigLike): string[] {
-  if (!providers) {
-    return [];
-  }
-
-  const providerList = Array.isArray(providers) ? providers : [providers];
-  return providerList.flatMap((provider) => {
-    if (typeof provider === 'string') {
-      return [provider];
-    }
-    if (!isRecord(provider)) {
-      return [];
-    }
-    if (typeof provider.id === 'string') {
-      return [provider.id];
-    }
-
-    try {
-      return getProviderIds([provider]);
-    } catch {
-      return [];
-    }
-  });
-}
-
-export function getCloudTargetIdFromProviders(providers: ProviderConfigLike): string | undefined {
-  if (!providers) {
-    return undefined;
-  }
-
-  const providerList = Array.isArray(providers) ? providers : [providers];
-  for (const provider of providerList) {
-    const directCloudTargetId = getCloudTargetIdFromProviderId(provider);
-    if (directCloudTargetId) {
-      return directCloudTargetId;
-    }
-    if (!isRecord(provider)) {
-      continue;
-    }
-
-    const linkedCloudTargetId = getLinkedCloudTargetId(provider);
-    if (linkedCloudTargetId) {
-      return linkedCloudTargetId;
-    }
-
-    const idCloudTargetId = getCloudTargetIdFromProviderId(provider.id);
-    if (idCloudTargetId) {
-      return idCloudTargetId;
-    }
-
-    for (const [providerId, providerConfig] of Object.entries(provider)) {
-      const keyCloudTargetId = getCloudTargetIdFromProviderId(providerId);
-      if (keyCloudTargetId) {
-        return keyCloudTargetId;
-      }
-
-      const mappedLinkedCloudTargetId = getLinkedCloudTargetId(providerConfig);
-      if (mappedLinkedCloudTargetId) {
-        return mappedLinkedCloudTargetId;
-      }
-    }
-  }
-
-  return undefined;
-}
-
-export function getRedteamGenerationContextFromProviders(
-  providers: ProviderConfigLike,
-): RedteamGenerationContext {
-  const providerTargetIds = getProviderTargetIds(providers);
-  return {
-    providerTargetIds,
-    cloudTargetId:
-      getCloudTargetIdFromProviders(providers) ?? getCloudTargetIdFromTargetIds(providerTargetIds),
-  };
 }
 
 export function resolveRedteamGenerationContext({
